refactor(api): rename and move files under organization folder
